Clean up rust directory structure
