+++ .vscode/settings.json
@@ -1,8 +1,11 @@
 {
-    "editor.rulers": [100],
+    "editor.rulers": [
+        100
+    ],
     "eslint.autoFixOnSave": true,
     "files.insertFinalNewline": true,
     "editor.codeActionsOnSave": {
         "source.fixAll.eslint": true
-    }
+    },
+    "rust-client.enableMultiProjectSetup": true
 }
